Forming a Magic Square
We define a magic square to be an n x n matrix of distinct positive
integers from 1 to n-square  where the sum of any row, column, or
diagonal (of length n) is always equal to the same number (i.e., the
magic constant).

Consider a 3x3 matrix , s, of integers in the inclusive range[1,9] . We
can convert any digit ,a, to any other digit, b, in the range[1,9]  at
cost |a-b|.

Given s, convert it into a magic square at minimal cost by changing
zero or more of its digits. Then print this cost on a new line.