Viral Advertising
Given an integer, n, find and print the total number of people who
liked HackerLand Enterprise's advertisement during the first n days. It
is guaranteed that no two people have any friends in common and, after
a person shares the advertisement with a friend, the friend always sees
it the next day.