Beautiful Days at the Movies
Lily likes to play games with integers and their reversals. For some
integer x, we define reversal(x) to be the reversal of all digits in x.
Logan wants to go to the movies with Lily on some day x satisfying
i<=x<=j, but he knows she only goes to the movies on days she considers
to be beautiful. Lily considers a day to be beautiful if the absolute
value of the difference between x and reversed(x) is evenly divisible
by k.

Given i, j, and k, count and print the number of beautiful days when
Logan and Lily can go to the movies.